Added new features for the admin panel and edited the admin panel UI
Categories can now be listed in the admin panel.
Added pagination feature for category list in admin panel.
Now a new category can be added from the admin panel.
Now the Blog List can be downloaded in Excel file format from the admin panel.
UI update for widgets section in admin panel.
Now the statistics of the website can be viewed in the widgets section in the admin panel.
Now admin information can be viewed through the admin panel.
Added Google Charts for category statistics to the admin panel.